Function to import, save and overwrite PTA Disbursement level data to DuckDB Database

Reads PTA Disbursement level data from a CSV file, processes it, cleans the column names, and then saves the data to a DuckDB database. The existing data in the "pta_disbursement_level" table in the database will be overwritten.

Usage

create_pta_disbursement_level_to_db(input_csv)

Arguments

input_csv

A string. Path to the input CSV file containing the PTA data.

Details

Before running this function, it is recommended that users first inspect the data by running the `read_pta_disbursement_level_from_csv()` function to ensure that the data has the expected structure and column types.

The function checks if both the CSV file and the DuckDB database file exist. If either file is not accessible, the function will stop and return an error message. It uses internal `noradstats` functions to read and process the data. The `janitor` package is used to clean column names. The resulting data is then written to the DuckDB database, overwriting the existing "pta_disbursement_level" table.

Examples

if (FALSE) { # \dontrun{
# Example usage:
# Path to the CSV file containing the PTA data
input_csv <- "path/to/your_pta_report.csv"

# Call the function to save and overwrite the Statsys data
create_pta_disbursement_level_to_db(input_csv)
} # }